+++ trunk/src/app/srs_app_process.hpp
@@ -53,6 +53,9 @@ public:
     // @param argv the argv for binary path, the argv[0] generally is the binary.
     // @remark the argv[0] must be the binary.
     virtual srs_error_t initialize(std::string binary, std::vector<std::string> argv);
+private:
+    // Redirect standard I/O.
+    virtual srs_error_t redirect_io();
 public:
     // Start the process, ignore when already started.
     virtual srs_error_t start();