+++ trunk/src/kernel/srs_kernel_codec.cpp
@@ -806,41 +806,47 @@ srs_error_t SrsFormat::avc_demux_sps_pps(SrsBuffer* stream)
     }
     int8_t numOfSequenceParameterSets = stream->read_1bytes();
     numOfSequenceParameterSets &= 0x1f;
-    if (numOfSequenceParameterSets != 1) {
+    if (numOfSequenceParameterSets < 1) {
         return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode SPS");
     }
-    if (!stream->require(2)) {
-        return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode SPS size");
+    // Support for multiple SPS, then pick the first non-empty one.
+    for (int i = 0; i < numOfSequenceParameterSets; ++i) {
+        if (!stream->require(2)) {
+            return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode SPS size");
+        }
+        uint16_t sequenceParameterSetLength = stream->read_2bytes();
+        if (!stream->require(sequenceParameterSetLength)) {
+            return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode SPS data");
+        }
+        if (sequenceParameterSetLength > 0) {
+            vcodec->sequenceParameterSetNALUnit.resize(sequenceParameterSetLength);
+            stream->read_bytes(&vcodec->sequenceParameterSetNALUnit[0], sequenceParameterSetLength);
+        }
     }
-    uint16_t sequenceParameterSetLength = stream->read_2bytes();
-    if (!stream->require(sequenceParameterSetLength)) {
-        return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode SPS data");
-    }
-    if (sequenceParameterSetLength > 0) {
-        vcodec->sequenceParameterSetNALUnit.resize(sequenceParameterSetLength);
-        stream->read_bytes(&vcodec->sequenceParameterSetNALUnit[0], sequenceParameterSetLength);
-    }
+
     // 1 pps
     if (!stream->require(1)) {
         return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode PPS");
     }
     int8_t numOfPictureParameterSets = stream->read_1bytes();
     numOfPictureParameterSets &= 0x1f;
-    if (numOfPictureParameterSets != 1) {
-        return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode PPS");
+    if (numOfPictureParameterSets < 1) {
+        return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode SPS");
     }
-    if (!stream->require(2)) {
-        return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode PPS size");
+    // Support for multiple PPS, then pick the first non-empty one.
+    for (int i = 0; i < numOfPictureParameterSets; ++i) {
+        if (!stream->require(2)) {
+            return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode PPS size");
+        }
+        uint16_t pictureParameterSetLength = stream->read_2bytes();
+        if (!stream->require(pictureParameterSetLength)) {
+            return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode PPS data");
+        }
+        if (pictureParameterSetLength > 0) {
+            vcodec->pictureParameterSetNALUnit.resize(pictureParameterSetLength);
+            stream->read_bytes(&vcodec->pictureParameterSetNALUnit[0], pictureParameterSetLength);
+        }
     }
-    uint16_t pictureParameterSetLength = stream->read_2bytes();
-    if (!stream->require(pictureParameterSetLength)) {
-        return srs_error_new(ERROR_HLS_DECODE_ERROR, "decode PPS data");
-    }
-    if (pictureParameterSetLength > 0) {
-        vcodec->pictureParameterSetNALUnit.resize(pictureParameterSetLength);
-        stream->read_bytes(&vcodec->pictureParameterSetNALUnit[0], pictureParameterSetLength);
-    }
-    
     return avc_demux_sps();
 }
 
@@ -1064,10 +1070,57 @@ srs_error_t SrsFormat::avc_demux_sps_rbsp(char* rbsp, int nb_rbsp)
     if ((err = srs_avc_nalu_read_uev(&bs, pic_height_in_map_units_minus1)) != srs_success) {
         return srs_error_wrap(err, "read pic_height_in_map_units_minus1");;
     }
-    
-    vcodec->width = (int)(pic_width_in_mbs_minus1 + 1) * 16;
-    vcodec->height = (int)(pic_height_in_map_units_minus1 + 1) * 16;
-    
+
+    int8_t frame_mbs_only_flag = -1;
+    if ((err = srs_avc_nalu_read_bit(&bs, frame_mbs_only_flag)) != srs_success) {
+        return srs_error_wrap(err, "read frame_mbs_only_flag");;
+    }
+    if(!frame_mbs_only_flag) {
+        /* Skip mb_adaptive_frame_field_flag */
+        int8_t mb_adaptive_frame_field_flag = -1;
+        if ((err = srs_avc_nalu_read_bit(&bs, mb_adaptive_frame_field_flag)) != srs_success) {
+            return srs_error_wrap(err, "read mb_adaptive_frame_field_flag");;
+        }
+    }
+
+    /* Skip direct_8x8_inference_flag */
+    int8_t direct_8x8_inference_flag = -1;
+    if ((err = srs_avc_nalu_read_bit(&bs, direct_8x8_inference_flag)) != srs_success) {
+        return srs_error_wrap(err, "read direct_8x8_inference_flag");;
+    }
+
+    /* We need the following value to evaluate offsets, if any */
+    int8_t frame_cropping_flag = -1;
+    if ((err = srs_avc_nalu_read_bit(&bs, frame_cropping_flag)) != srs_success) {
+        return srs_error_wrap(err, "read frame_cropping_flag");;
+    }
+    int32_t frame_crop_left_offset = 0, frame_crop_right_offset = 0,
+            frame_crop_top_offset = 0, frame_crop_bottom_offset = 0;
+    if(frame_cropping_flag) {
+        if ((err = srs_avc_nalu_read_uev(&bs, frame_crop_left_offset)) != srs_success) {
+            return srs_error_wrap(err, "read frame_crop_left_offset");;
+        }
+        if ((err = srs_avc_nalu_read_uev(&bs, frame_crop_right_offset)) != srs_success) {
+            return srs_error_wrap(err, "read frame_crop_right_offset");;
+        }
+        if ((err = srs_avc_nalu_read_uev(&bs, frame_crop_top_offset)) != srs_success) {
+            return srs_error_wrap(err, "read frame_crop_top_offset");;
+        }
+        if ((err = srs_avc_nalu_read_uev(&bs, frame_crop_bottom_offset)) != srs_success) {
+            return srs_error_wrap(err, "read frame_crop_bottom_offset");;
+        }
+    }
+
+    /* Skip vui_parameters_present_flag */
+    int8_t vui_parameters_present_flag = -1;
+    if ((err = srs_avc_nalu_read_bit(&bs, vui_parameters_present_flag)) != srs_success) {
+        return srs_error_wrap(err, "read vui_parameters_present_flag");;
+    }
+
+    vcodec->width = ((pic_width_in_mbs_minus1 + 1) * 16) - frame_crop_left_offset * 2 - frame_crop_right_offset * 2;
+    vcodec->height = ((2 - frame_mbs_only_flag) * (pic_height_in_map_units_minus1 + 1) * 16) \
+                    - (frame_crop_top_offset * 2) - (frame_crop_bottom_offset * 2);
+
     return err;
 }